module/video/requirements.txt
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
mtcnn==0.0.9
numpy==1.16.4
opencv-contrib-python==4.1.2.30
opencv-python==4.1.0.25
scikit-learn==0.21.3
scipy==1.3.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.12.0
flask
